DIGITAL_TRANSFORMATION_PROJECT/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.1
orjson==3.9.7
psycopg2==2.9.9
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.6
pypdf2==3.0.1
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2